pyproject.toml
[tool.poetry]
name = "project4"
version = "0.1.0"
description = ""
authors = ["hantaek <lht92229@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11, <3.13"
python-dotenv = "1.0.1"
langchain = "0.3.3"
langchain-openai = "0.2.2"
langchain-community = "0.3.2"
gradio = "5.0.2"
crewai = "0.76.9"
crewai-tools = "0.13.4"
tavily-python = "0.5.0"
duckduckgo-search = "6.3.3"
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
